getbits, printbits: consistently use unsigned type
Also, getbits() correctly accepts p==0

+++ printbits.c
@@ -5,7 +5,7 @@
  * print binary representation of the n least
  * significant bits of x. 
  */
-void printbits(long x, unsigned n)
+void printbits(unsigned x, int n)
 {
 	for (; n>0; n--)
 		putchar(x & 1 << n-1 ? '1' : '0');
